aci_bd_subnet: Improve support for multiple scope options
This patch supports combinations of private/public with shared as values
for the scope parameter. The input now is defined as a list, which means
Ansible will accept either a comma-separated list, or a proper JSON/YAML
list definition.

The code ensures that private and public are mutually exclusive, and it
also checks for unknown options.